requirements.txt
imbalanced_learn==0.9.0
imblearn==0.0
keras==2.8.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
scikit_learn==1.0.2
scipy==1.8.0
tensorflow==2.8.0
biopython==1.79
